+++ Dessin.h
@@ -13,16 +13,16 @@
  ****************************************************************/
 
 // Trace un point à partir de coordonnées
-void trace_point(double x, double y, double red, double green, double blue, double size);
+void trace_point(float x, float y, float red, float green, float blue, float size);
 
 // Trace un point à partir d'un objet Point
-void trace_point(Point p, double red, double green, double blue, double size);
+void trace_point(Point p, float red, float green, float blue, float size);
 
 // Trace un segement à partir des coordonnées des points
-void trace_segment(double x0, double y0, double x1, double y1, double red, double green, double blue, double size);
+void trace_segment(float x0, float y0, float x1, float y1, float red, float green, float blue, float size);
 
 // Trace un segment à partir de deux objets points
-void trace_segment(Point p0, Point p1, double red, double green, double blue, double size);
+void trace_segment(Point p0, Point p1, float red, float green, float blue, float size);
 
 // Trace un segment à partir d'un objet segment
-void trace_segment(Segment s, double red, double green, double blue, double size);
+void trace_segment(Segment s, float red, float green, float blue, float size);
