+++ PointMassique.h
@@ -9,13 +9,13 @@ class PointMassique : public Point
   public:
 
     PointMassique();
-    PointMassique(const Point &p, double W);
-    PointMassique(double X, double Y, double W);
+    PointMassique(const Point &p, float W);
+    PointMassique(float X, float Y, float W);
 
     PointMassique &operator+=(PointMassique const &p);
-    PointMassique &operator*=(double scalaire);
+    PointMassique &operator*=(float scalaire);
 
-    double getW() const;
+    float getW() const;
 
     void display() const;
 
@@ -24,11 +24,11 @@ class PointMassique : public Point
 
   private:
 
-    double w;
+    float w;
 
 };
 
 PointMassique operator+(PointMassique const &p1, PointMassique const &p2);
-PointMassique operator*(PointMassique const &p, double scalaire);
+PointMassique operator*(PointMassique const &p, float scalaire);
 
 #endif
